Creada base microservicio reservas

+++ Evaluacion 03/Entregable/MicroServicios/reservas/src/main/java/Microservicio/repository/HotelRepository.java
@@ -3,6 +3,6 @@ package Microservicio.repository;
 import Microservicio.entity.Hotel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HotelRepository extends JpaRepository<Hotel,Long> {
+public interface HotelRepository extends JpaRepository<Hotel,Integer> {
 
 }
