+++ Evaluacion 03/Entregable/MicroServicios/reservas/src/main/java/Microservicio/repository/ReservaRepository.java
@@ -3,6 +3,6 @@ package Microservicio.repository;
 import Microservicio.entity.Reserva;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReservaRepository extends JpaRepository<Reserva,Long> {
+public interface ReservaRepository extends JpaRepository<Reserva,Integer> {
 
 }
